fix(next): move ingest validator export to lib to satisfy route export rules
